Remove external dependencies from AWS integration test
